requirements.txt
transformers
pillow
torch
torchvisio
gradio
pillow


